Use a markdown cell for error flagging, prettier than %%html
Markdown is (mostly) a superset of HTML, meaning there's a lot of HTML one can include directly as
markdown content. Papermill can use this to have a slightly cleaner error alert at the top of a
failed notebook, by not having both code and output, just output only.

This makes the rendering cleaner both in an interactive Jupyter instance, as well as on the GitHub
and NBViewer static renderers:

Before:
- <https://gist.github.com/huonw/b08434842b6b9f6d45090141ae4327dd>
- <https://nbviewer.jupyter.org/gist/huonw/b08434842b6b9f6d45090141ae4327dd>

After:
- <https://gist.github.com/huonw/f22a6caf6bf8c31538000b88e28c05da>
- <https://nbviewer.jupyter.org/gist/huonw/f22a6caf6bf8c31538000b88e28c05da>