Include link to the cell with an error, in failure heading
Papermill currently inserts an error message "An Exception was encountered at 'In [NNN]'" at the top
of a notebook that fails to execute, where NNN is the cell execution count of the failing cell. This
patch expands the "In [NNN]" text to be a link to the relevant cell, to make it easy to jump
to. This works by inserting another markdown cell immediately before the failing cell, where the
markdown cell contains an HTML anchor (specifically, `id="papermill-error-cell"`). The heading then
links to it with a `#papermill-error-cell` fragment HREF.

This helps reducing the scrolling and searching required for long notebooks where the error cell
might be one of dozens, or hidden among long output. For an example of the latter, see
<https://nbviewer.jupyter.org/gist/huonw/dd13fa49b5876bfea26b69832702d573>.