Bugfix/pickle errors (#629)
* Add tests for showing unpickling error

* Add fix so exceptions can be loaded from pickle

This fixes an issue where the PapermillExecutionError can be pickled but will
not be unpicklable.

Pythons BaseException class has a handling to make sure that exceptions are able
to be loaded once they are pickled. This works by ensuring the args used to generate the instance
are saved on the exception. However by only passing the message to the superclass, the pickle
module was unable to load a new instance of PapermillExecutionError from a file.

A new __str__ method was added to preserve the functionality of calling str() on the exception and
only getting the error message back, while ensuring the superclass had access to all instantiating
arguments.

Co-authored-by: Tyler Calder <calder-ty@protonmail.com>
Co-authored-by: Matthew Seal <mseal007@gmail.com>