Drop the dependency on ansicolors (#822)
* Drop the dependency on ansicolors

The package is not actively maintained (last released in 2017), and the
function we use is just based on a single simple regular expression.
Copy the regular expression into papermill and drop the external
dependency.

* Explicitly precompile the ANSI-colors regular expression

This may not be terribly important, as the re module caches patterns,
but it doesn’t hurt. Corresponds to
https://github.com/jonathaneunice/colors/pull/2.

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>